Simplify propagation
The OpenTracing spec mandates that the carrier type be determined at
runtime, for the sake of extensibility (obviously, this may also _fail_
at runtime, if the carrier type is not supported). A safer approach is
to encode the required carrier formats using a typeclass, yet have the
use site choose the desired carrier by supplying the appropriate
`Prism`. It's still possible to support other carrier types, and use of
an unsupported one is a compile-time error.