Fix Zipkin BinaryAnnotation value encoding
The Thrift IDL says:

> Serialized thrift bytes, in TBinaryProtocol format.
>
> For legacy reasons, byte order is big-endian. See THRIFT-3217.

The truth is, they just use a ByteBuffer. It's not even clear where they
set/check the byte order, so this may need to be revisited again in the
future.

At least, the Zipkin Thrift format is legacy, so we may even drop
support for it at some point.