Drop stack, hpack, and cloudtrace
stack is not in a good shape it seems, it fails to install on some
platforms, and it's not possible to boostrap it from cabal-install. Thus
move to cabal-install, and drop hpack in the process.

Due to a lack of time and infrastructure to update to the v2 cloudtrace
API and a recent gogol, drop opentracing-cloudtrace as well.